chore: set kubectl flags on root command
This changeset compliments #454 by making the cli flags more
readable in documentation and usage help.

The options AddKubectlFlags is only applied to the the root
argo rollouts command and has been removed from all the child
commands. The result of this from a documentation perspective
is options are now split between options specific to the command
and options inherited from parent.

One side effect of this is tests that use inherited args need to call
`o.AddKubectlFlags(cmd)` explicitly.