Use numeric user in Dockerfile
This allows the use of a security context in kubernetes with
runAsNonRoot: true, because kubernetes can inspect the image and see
that the USER is != 0.

Removed user and group creation, it isn't necessary to get a working
docker image. Also when running in openshift, the user in the docker
image is usually ignored and a random user is used to run the image.