fix: fetch secrets on-demand to fix controller boot for large clusters
Using a list cache of secrets is not needed as we do not need to react
to any changes and we can always fetch the individual secrets we need
for analysis arguments when the time comes. For large kubernetes
clusters with thousands of secrets this causes the argo-rollouts
controller to be unbootable.

Co-authored-by: Jason Yanowitz <jason.yanowitz@getbraintree.com>
Signed-off-by: Eric Mrak <emrak@paypal.com>