fix: set gopath in makefile
We currently assume that GOPATH is set within the developers environment
go has some defaults that get used when no GOPATH is defined. If we use
`go env GOPATH` to set the env var for the Makefile we will then also fall
back to golang defaults which help contributior experiance.

Signed-off-by: zachaller <zachaller@users.noreply.github.com>