fix: add configurable subset level DestinationRules (#4332)
* fix: add configurable subset level DestinationRules

The code here is specifically to address #3125, however, a commit previously (#3126) caused a bug (#3141). After more testing, the bug is still reproducible with the changed in commit #3126 applied to todays codebase, so my proposed solution for both is to ensure that patches whom have subset rule names do not get deleted.

This fixes the regression, as the code within #3126 introduced routes from the managed routes section that did not add up too 100% total weight, and also ensures that subset destination rules do not get deleted.

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>

* fix: add additional checks to not overwrite subset DestinalRule weights

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>

* feat: changed approach, using patches

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>

* chore(docs): described how to use additional subset DestinationRules

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>

* typo

Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>

---------

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>
Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>
Co-authored-by: Zach Aller <zachaller@users.noreply.github.com>