fix(trafficrouting): Fix rollback behavior for canary with trafficrouting and .DynamicStableScale=true (#4035)
* fix(trafficrouting): Fix rollback behavior for canary with trafficrouting and .DynamicStableScale=true

Without the fix, controller always scales stable up to 100%, while dynamically scaling canary down.
If rollback or abort occurs on later steps of rollout, it can cause various issues, e.g. surge of hundreds of new pods.

This fix ensures that when a rollout is aborted and .DynamicStableScale is enabled,
the StableRS dynamically scales up to 100% as NewRS scales down based on steps in reverse order.

Signed-off-by: Armen Shakhbazian <armen.shakhbazian@gmail.com>

* test(controller): remove code duplication in replicaset/canary tests

Signed-off-by: Armen Shakhbazian <armen.shakhbazian@gmail.com>

* lint updates

Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>

---------

Signed-off-by: Armen Shakhbazian <armen.shakhbazian@gmail.com>
Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>
Co-authored-by: Zach Aller <zachaller@users.noreply.github.com>