fix: ReplicaSetReferenced does not check istio DestinationRules
The isReplicaSetReferenced function checks:
- Static references in rollout status (StableRS, CurrentPodHash, Weights.Canary.PodTemplateHash, Weights.Stable.PodTemplateHash)
- Service selectors for Canary/BlueGreen services

However, when using Istio with subset-level traffic splitting (DestinationRules),
the function doesn't check if the ReplicaSet's pod template hash is still
referenced in the Istio DestinationRule subsets.
This means a ReplicaSet can be scaled down even though Istio is still routing
traffic to that subset.

When using Istio with DestinationRule subsets (without explicit canary/stable services), the isReplicaSetReferenced function doesn't check the DestinationRule to see if the pod template hash is still referenced in a subset.

The problem is:
- The rollout status (e.g., Weights.Canary.PodTemplateHash) gets updated to the new hash
- But the Istio DestinationRule subsets still have the old hash because UpdateHash hasn't been called yet (or returns early due to checkReplicasAvailable)
- The ReplicaSet with the old hash gets scaled down because isReplicaSetReferenced returns false
- Traffic is still being routed to that subset, causing 503s

Signed-off-by: Andrew Brown <andrew.brown@wealthsimple.com>